tests/wpt/tests/content-security-policy/reporting/report-strips-username-password.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/content-security-policy/support/testharness-helper.js"></script>
  <meta http-equiv="Content-Security-Policy" content="img-src 'none'">
</head>
<body>
  <script>
  async_test(t => {
    addEventListener("securitypolicyviolation", t.step_func_done(e => {
      assert_equals(e.blockedURI, "https://evil.com/img.png");
    }));

    var i = document.createElement("img");
    i.src = "https://user:password@evil.com/img.png";
  }, "Reported blockedURI does not contain username or password.");
  </script>
</body>
